requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
django-bootstrap3==15.0.0
django-paypal==1.1.1
idna==2.10
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
